+++ src/main/java/com/achobeta/domain/email/service/impl/EmailServiceImpl.java
@@ -4,7 +4,7 @@ import com.achobeta.domain.email.model.po.EmailMessage;
 import com.achobeta.domain.email.model.vo.VerificationCodeTemplate;
 import com.achobeta.domain.email.service.EmailSender;
 import com.achobeta.domain.email.service.EmailService;
-import com.achobeta.domain.html.service.HtmlEngine;
+import com.achobeta.domain.templateengine.service.HtmlEngine;
 import com.achobeta.exception.GlobalServiceException;
 import com.achobeta.redis.RedisCache;
 import lombok.RequiredArgsConstructor;
