+++ src/main/java/com/achobeta/domain/templateengine/model/po/MarkdownResource.java
@@ -1,4 +1,4 @@
-package com.achobeta.domain.html.model.po;
+package com.achobeta.domain.templateengine.model.po;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,16 +8,16 @@ import lombok.NoArgsConstructor;
  * Created With Intellij IDEA
  * Description:
  * User: 马拉圈
- * Date: 2024-09-05
- * Time: 9:00
+ * Date: 2024-09-07
+ * Time: 16:39
  */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class HtmlReplaceResource {
+public class MarkdownResource {
 
-    private String target;
+    private String template;
 
-    private String html;
+    private Object context;
 
 }
